# perf(ext/node): improve createHash performance