--- a/ext/node/polyfills/internal/crypto/hash.ts
+++ b/ext/node/polyfills/internal/crypto/hash.ts
@@ -1,10 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 // Copyright Joyent, Inc. and Node.js contributors. All rights reserved. MIT license.
 
-import {
-  TextDecoder,
-  TextEncoder,
-} from "internal:deno_web/08_text_encoding.js";
+import { TextEncoder } from "internal:deno_web/08_text_encoding.js";
 import { Buffer } from "internal:deno_node/buffer.ts";
 import { Transform } from "internal:deno_node/stream.ts";
 import {